pyproject.toml
[tool.poetry]
name = "voice-conversation"
version = "0.1.0"
description = "A voice-based conversation assistant using speech recognition and text-to-speech"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
openai = "^1.3.0"
websockets = "^11.0.3"
aiofiles = "^23.2.1"
requests = "^2.31.0"
pyinstaller = "^6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
voice-assistant = "main:main"
